Added modules and data
a) Textured and coloured UI interface, did all that using Photoshop to every panel and UI graphic.

b) In-game guide module (can be customized with images, a contextual menu and text).

c) Rc-Sound module. It changes the song that the client plays based on areas. The current track and b e seen on Sound Info panel.  It has Final Fantasy X tracks sampled to be an infinite loop with ambience (people noise, time stations, etc.) (made on Adobe Audition).

d) OTC-bot based on Lunia's bot.

e) Fixed battle window for 8.60 client.

f) Researched everything to make it work the best as possible with 8.6.

g) Optimized action bar (without icons for 8.6).

h) Oen's module for location change (it shows a square with the name of the area).